requirements.txt
amqp==2.5.0
anyjson==0.3.3
appnope==0.1.0
beautifulsoup4==4.8.0
behave-django==1.4.0
billiard==3.6.0.0
boto3==1.15.9
botocore==1.18.9
celery==4.3.0
certifi==2020.6.20
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.14
django-ajax-selects==1.9.1
django-autocomplete-light==3.4.1
django-betterforms==1.2
django-bootstrap-pagination==1.7.1
django-bootstrap3==11.1.0
django-ckeditor==5.9.0
django-extensions==3.0.2
django-extra-views==0.13.0
django-filter==2.3.0
django-form-utils==1.0.3
django-image-cropping==1.5.0
django-picklefield==2.0
django-recaptcha2==1.4.1
django-rest-swagger==2.2.0
django-reversion==3.0.4
djangorestframework==3.11.0
django-storages==1.10.1
easy-thumbnails==2.7.0
Faker==4.1.2
freezegun==0.3.15
git+https://github.com/danpalmer/django-google-charts.git
git+https://github.com/fenrrir/csv2py.git
git+https://github.com/jedie/django-reversion-compare.git
git+https://gitbot:n9HPkohCgusyFuHmBFds@gitlab.ifpb.edu.br/cdgs/ifpb-django-importutils.git@f9dcd8abb3b4defc501740c46ead101eb36cd81d#egg=ifpb_django_importutils
git+https://gitbot:n9HPkohCgusyFuHmBFds@gitlab.ifpb.edu.br/cdgs/ifpb-django-menu.git@303861631775fdcfb1702f2f2053d2c1ba51ca28#egg=ifpb_django_menu
git+https://gitbot:n9HPkohCgusyFuHmBFds@gitlab.ifpb.edu.br/cdgs/ifpb-django-permissions.git@a9d187ca3af6c909d160a0b4d0d980bf1106aa1e#egg=ifpb_django_permissions
#gnureadline==8.0.0
gunicorn==20.0.4
kombu==4.6.3
minio==5.0.10
model-mommy==2.0.0
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2-binary==2.8.5
Pygments==2.6.1
python-memcached==1.59
raven==6.10.0
redis==3.3.4
reportlab==3.5.23
requests==2.22.0
selenium==3.141.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
splinter==0.13.0
sqlparse==0.3.0
git+https://gitbot:n9HPkohCgusyFuHmBFds@gitlab.ifpb.edu.br/cdgs/suaprest.git@v0.2
git+https://gitbot:n9HPkohCgusyFuHmBFds@gitlab.ifpb.edu.br/dgti/jaiminho-client.git@v0.2
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
xlsxwriter==1.2.9
